backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
pydantic
requests
beautifulsoup4
redis
stripe
supabase
yfinance
litellm
